pyproject.toml
[tool.poetry]
name = "abd-weppapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
matplotlib = ">=3.3.4,"
pandas = ">=1.2.4,"
plotly = ">=4.14.3,"
psycopg2 = ">=2.8.6,"
pytz = ">=2021.1,"
scikit-learn = ">=1.0.2,"
scipy = ">=1.6.2,"
seaborn = ">=0.11.1,"
guardian = "^0.2.2"
crispy-bootstrap5 = "==0.6"

[tool.poetry.group.django.dependencies]
Django = "==4.0.4"
django-crispy-forms = "^1.11.2,"
django-timescaledb = "==0.2.11"
django-extensions = "==3.1.5"
django-guardian = "^2.4.0"
djangorestframework = ">=3.14.0"
django-cors-headers = "^4.3.1"
django-filter = "^23.0"

[tool.poetry.group.pytables.dependencies]
cython = "^3.0.8"
numexpr = "^2.9.0"
packaging = "^23.2"
py-cpuinfo = "^9.0.0"
tables = "^3.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
